+++ Yao/YaoEvalWire.cpp
@@ -250,7 +250,7 @@ void YaoEvalWire::convcbit2s(GC::Processor<whole_type>& processor,
 	for (int i = 0; i < DIV_CEIL(instruction.get_n(), unit); i++)
 	{
 		auto& dest = processor.S[instruction.get_r(0) + i];
-		dest.resize_regs(min(unsigned(unit), instruction.get_n() - i * unit));
+		dest.resize_regs(min(size_t(unit), instruction.get_n() - i * unit));
 		for (auto& reg : dest.get_regs())
 			reg.set(0);
 	}
