+++ Yao/YaoGarbleWire.cpp
@@ -238,7 +238,7 @@ void YaoGarbleWire::convcbit2s(GC::Processor<whole_type>& processor,
 	for (int i = 0; i < DIV_CEIL(instruction.get_n(), unit); i++)
 	{
 		auto& dest = processor.S[instruction.get_r(0) + i];
-		int n = min(unsigned(unit), instruction.get_n() - i * unit);
+		int n = min(size_t(unit), instruction.get_n() - i * unit);
 		dest.resize_regs(n);
 		for (int j = 0; j < n; j++)
 			dest.get_reg(j).public_input(
